Fixes #2301 Organize `extensionspmaxi-block` folder
