+++ src/blocks/button-maxi/edit.js
@@ -11,7 +11,8 @@ import { RawHTML, createRef, forwardRef, useEffect } from '@wordpress/element';
  * Internal dependencies
  */
 import Inspector from './inspector';
-import { MaxiBlockComponent, Toolbar } from '../../components';
+import { MaxiBlockComponent } from '../../extensions/maxi-block';
+import { Toolbar } from '../../components';
 import MaxiBlock, {
 	getMaxiBlockBlockAttributes,
 } from '../../components/maxi-block';
