+++ src/blocks/map-maxi/edit.js
@@ -9,7 +9,8 @@ import { withSelect } from '@wordpress/data';
  * Internal dependencies
  */
 import Inspector from './inspector';
-import { MaxiBlockComponent, Toolbar } from '../../components';
+import { MaxiBlockComponent } from '../../extensions/maxi-block';
+import { Toolbar } from '../../components';
 import MaxiBlock, {
 	getMaxiBlockBlockAttributes,
 } from '../../components/maxi-block';
