+++ src/blocks/text-maxi/edit.js
@@ -16,7 +16,8 @@ import {
  * Internal dependencies
  */
 import Inspector from './inspector';
-import { MaxiBlockComponent, Toolbar } from '../../components';
+import { MaxiBlockComponent } from '../../extensions/maxi-block';
+import { Toolbar } from '../../components';
 import MaxiBlock, {
 	getMaxiBlockBlockAttributes,
 } from '../../components/maxi-block';
