+++ src/components/index.js
@@ -39,7 +39,6 @@ export { default as Indicators } from './indicators';
 export { default as InfoBox } from './info-box';
 export { default as InnerBlocks } from './inner-blocks';
 export { default as MapControl } from './map-control';
-export { default as MaxiBlockComponent } from '../extensions/maxi-block';
 export { default as MotionControl } from './motion-control';
 export { default as MotionPreview } from './motion-preview';
 export { default as NumberCounterControl } from './number-counter-control';
